+++ tests/performance_profiling/canvas_performance_profiling.txt
@@ -1,27 +1,29 @@
 _____________________________________________________________________________________________________________________________________________
 Performance Profiling for free_draw()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 19:40:07  Samples:  463
- /_//_/// /_\ / //_// / //_'/ //     Duration: 0.526     CPU time: 0.016
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:45:58  Samples:  2
+ /_//_/// /_\ / //_// / //_'/ //     Duration: 0.003     CPU time: 0.000
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_canvas.py:19
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_canvas.py:19
 
-0.525 test_performance_profile_free_draw  test_canvas.py:18
-`- 0.525 free_draw  pages\canvas.py:5
-   `- 0.525 st_canvas  streamlit_drawable_canvas\__init__.py:56
-         [88 frames hidden]  streamlit, pandas, pyarrow, inspect, ...
+0.002 test_performance_profile_free_draw  performance_profile_canvas.py:18
+`- 0.002 free_draw  pages\canvas.py:5
+   `- 0.002 st_canvas  streamlit_drawable_canvas\__init__.py:56
+         [19 frames hidden]  streamlit, logging
+            0.001 TextIOWrapper.write  <built-in>
+            0.001 LogRecord.__init__  logging\__init__.py:292
 
 
 
 _____________________________________________________________________________________________________________________________________________
 Performance Profiling for free_draw()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 19:40:08  Samples:  0
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:45:58  Samples:  0
  /_//_/// /_\ / //_// / //_'/ //     Duration: 0.000     CPU time: 0.000
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_canvas.py:35
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_canvas.py:35
 
 No samples were recorded.
