+++ tests/performance_profiling/ocr_performance_profiling.txt
@@ -1,57 +1,57 @@
 _____________________________________________________________________________________________________________________________________________
-Performance Profiling for ninja_ocr()
-_____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 19:54:07  Samples:  10
- /_//_/// /_\ / //_// / //_'/ //     Duration: 1.634     CPU time: 0.000
-/   _/                      v5.0.0
-
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_ocr.py:21
-
-1.623 test_performance_profile_ninja_ocr  test_ocr.py:20
-`- 1.623 ninja_OCR  ocr\ocr.py:24
-   `- 1.623 post  requests\api.py:103
-         [27 frames hidden]  requests, urllib3, http, socket, ssl,...
-            1.192 _SSLSocket.read  <built-in>
-
-
-
-_____________________________________________________________________________________________________________________________________________
 Performance Profiling for Gemini_Flash_OCR()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 19:54:09  Samples:  65
- /_//_/// /_\ / //_// / //_'/ //     Duration: 10.667    CPU time: 0.031
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:46:02  Samples:  67
+ /_//_/// /_\ / //_// / //_'/ //     Duration: 9.567     CPU time: 0.016
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_ocr.py:35
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_ocr.py:35
 
-10.664 test_performance_profile_Gemini_Flash_OCR  test_ocr.py:34
-`- 10.663 Gemini_Flash_OCR  ocr\ocr.py:36
-   |- 7.263 upload_file  google\generativeai\files.py:32
-   |     [64 frames hidden]  google, googleapiclient, httplib2, ht...
-   `- 3.400 GenerativeModel.generate_content  google\generativeai\generative_models.py:237
+9.563 test_performance_profile_Gemini_Flash_OCR  performance_profile_ocr.py:34
+`- 9.563 Gemini_Flash_OCR  ocr\ocr.py:36
+   |- 6.713 upload_file  google\generativeai\files.py:32
+   |     [70 frames hidden]  google, googleapiclient, httplib2, ht...
+   `- 2.850 GenerativeModel.generate_content  google\generativeai\generative_models.py:237
          [7 frames hidden]  google, grpc
-            3.397 _UnaryUnaryMultiCallable._blocking  grpc\_channel.py:1128
+            2.846 _UnaryUnaryMultiCallable._blocking  grpc\_channel.py:1128
 
 
 
 _____________________________________________________________________________________________________________________________________________
 Performance Profiling for Gemini_Pro_OCR()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 19:54:19  Samples:  10
- /_//_/// /_\ / //_// / //_'/ //     Duration: 5.391     CPU time: 0.000
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:46:12  Samples:  14
+ /_//_/// /_\ / //_// / //_'/ //     Duration: 4.330     CPU time: 0.031
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_ocr.py:49
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_ocr.py:49
 
-5.391 test_performance_profile_Gemini_Pro_OCR  test_ocr.py:48
-`- 5.391 Gemini_Pro_OCR  ocr\ocr.py:42
-   |- 3.847 upload_file  google\generativeai\files.py:32
+4.331 test_performance_profile_Gemini_Pro_OCR  performance_profile_ocr.py:48
+`- 4.330 Gemini_Pro_OCR  ocr\ocr.py:42
+   |- 2.777 upload_file  google\generativeai\files.py:32
    |     [29 frames hidden]  google, googleapiclient, httplib2, ht...
-   |        1.910 _SSLSocket.read  <built-in>
-   |        1.549 _SSLSocket.read  <built-in>
-   `- 1.543 GenerativeModel.generate_content  google\generativeai\generative_models.py:237
+   |        1.478 _SSLSocket.read  <built-in>
+   |        0.899 _SSLSocket.read  <built-in>
+   `- 1.553 GenerativeModel.generate_content  google\generativeai\generative_models.py:237
          [7 frames hidden]  google, grpc
-            1.542 _UnaryUnaryMultiCallable._blocking  grpc\_channel.py:1128
+            1.552 _UnaryUnaryMultiCallable._blocking  grpc\_channel.py:1128
+
+
+
+_____________________________________________________________________________________________________________________________________________
+Performance Profiling for ninja_ocr()
+_____________________________________________________________________________________________________________________________________________
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:46:16  Samples:  10
+ /_//_/// /_\ / //_// / //_'/ //     Duration: 1.433     CPU time: 0.000
+/   _/                      v5.0.0
+
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_ocr.py:21
+
+1.432 test_performance_profile_ninja_ocr  performance_profile_ocr.py:20
+`- 1.432 ninja_OCR  ocr\ocr.py:24
+   `- 1.432 post  requests\api.py:103
+         [27 frames hidden]  requests, urllib3, http, socket, ssl,...
+            0.954 _SSLSocket.read  <built-in>
