+++ tests/performance_profiling/performance_profile_canvas.py
@@ -1,15 +1,11 @@
 from pyinstrument import Profiler
-import bcrypt
-
 import os
 import sys
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
-
-from auth.hasher import Hasher
-
+from pages.canvas import free_draw, save_drawing 
 
 
-FILE = "hasher_performance_profiling.txt"
+FILE = "canvas_performance_profiling.txt"
 hr = '_____________________________________________________________________________________________________________________________________________'
 
 profiler =  Profiler()
@@ -19,36 +15,13 @@ with open(FILE, 'w') as file:
 file = open(FILE, "a") 
 
 
-# mock password list for testing.
-passwords = ['12345', 'abcdef', '%##@zxcvbn9087']
-
-# pick one password for testing from the above list
-password = '%##@zxcvbn9087'
-
-
-def test_performance_profile_hasher():
+def test_performance_profile_free_draw():
     profiler.start()
-    val = Hasher(passwords=passwords)
+    free_draw()
     profiler.stop()
 
     file.write(hr)
-    file.write(f"\nPerformance Profiling for Hasher()\n")
-    file.write(hr)
-    file.write(profiler.output_text())
-    file.write("\n\n")
-    
-    profiler.reset()
-
-
-def test_performance_profile_hasher_generate():
-    hasher = Hasher(passwords=password)
-
-    profiler.start()
-    hashed_pws = hasher.generate()
-    profiler.stop()
-
-    file.write(hr)
-    file.write(f"\nPerformance Profiling for Hasher()\n")
+    file.write(f"\nPerformance Profiling for free_draw()\n")
     file.write(hr)
     file.write(profiler.output_text())
     file.write("\n\n")
@@ -56,7 +29,20 @@ def test_performance_profile_hasher_generate():
     profiler.reset()
 
     
+def test_performance_profile_save_drawing():
+    can = free_draw()
+
+    profiler.start()
+    save_drawing(can)
+    profiler.stop()
+
+    file.write(hr)
+    file.write(f"\nPerformance Profiling for free_draw()\n")
+    file.write(hr)
+    file.write(profiler.output_text())
+    file.write("\n\n")
+    
+    profiler.reset()
 
 
-test_performance_profile_hasher()
-test_performance_profile_hasher_generate()
+
