+++ tests/performance_profiling/performance_profile_hasher.py
@@ -1,11 +1,15 @@
 from pyinstrument import Profiler
+import bcrypt
+
 import os
 import sys
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
-from pages.canvas import free_draw, save_drawing 
+
+from auth.hasher import Hasher
 
 
-FILE = "canvas_performance_profiling.txt"
+
+FILE = "hasher_performance_profiling.txt"
 hr = '_____________________________________________________________________________________________________________________________________________'
 
 profiler =  Profiler()
@@ -15,29 +19,20 @@ with open(FILE, 'w') as file:
 file = open(FILE, "a") 
 
 
-def test_performance_profile_free_draw():
+# mock password list for testing.
+passwords = ['12345', 'abcdef', '%##@zxcvbn9087']
+
+# pick one password for testing from the above list
+password = '%##@zxcvbn9087'
+
+
+def test_performance_profile_hasher():
     profiler.start()
-    free_draw()
+    val = Hasher(passwords=passwords)
     profiler.stop()
 
     file.write(hr)
-    file.write(f"\nPerformance Profiling for free_draw()\n")
-    file.write(hr)
-    file.write(profiler.output_text())
-    file.write("\n\n")
-    
-    profiler.reset()
-
-    
-def test_performance_profile_save_drawing():
-    can = free_draw()
-
-    profiler.start()
-    save_drawing(can)
-    profiler.stop()
-
-    file.write(hr)
-    file.write(f"\nPerformance Profiling for free_draw()\n")
+    file.write(f"\nPerformance Profiling for Hasher()\n")
     file.write(hr)
     file.write(profiler.output_text())
     file.write("\n\n")
@@ -45,7 +40,20 @@ def test_performance_profile_save_drawing():
     profiler.reset()
 
 
+def test_performance_profile_hasher_generate():
+    hasher = Hasher(passwords=password)
 
+    profiler.start()
+    hashed_pws = hasher.generate()
+    profiler.stop()
 
-test_performance_profile_free_draw()
-test_performance_profile_save_drawing()
+    file.write(hr)
+    file.write(f"\nPerformance Profiling for Hasher()\n")
+    file.write(hr)
+    file.write(profiler.output_text())
+    file.write("\n\n")
+    
+    profiler.reset()
+
+    
+
