+++ tests/performance_profiling/validator_performance_profiling.txt
@@ -1,11 +1,11 @@
 _____________________________________________________________________________________________________________________________________________
 Performance Profiling for Validator.validate_username()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:24:14  Samples:  0
- /_//_/// /_\ / //_// / //_'/ //     Duration: 0.000     CPU time: 0.000
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:47:34  Samples:  0
+ /_//_/// /_\ / //_// / //_'/ //     Duration: 0.001     CPU time: 0.000
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_validator.py:28
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_validator.py:28
 
 No samples were recorded.
 
@@ -14,11 +14,11 @@ No samples were recorded.
 _____________________________________________________________________________________________________________________________________________
 Performance Profiling for Validator.validate_email()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:24:14  Samples:  0
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:47:34  Samples:  0
  /_//_/// /_\ / //_// / //_'/ //     Duration: 0.000     CPU time: 0.000
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_validator.py:42
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_validator.py:42
 
 No samples were recorded.
 
@@ -27,11 +27,11 @@ No samples were recorded.
 _____________________________________________________________________________________________________________________________________________
 Performance Profiling for Validator.validate_name()
 _____________________________________________________________________________________________________________________________________________
-  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:24:14  Samples:  0
+  _     ._   __/__   _ _  _  _ _/_   Recorded: 20:47:34  Samples:  0
  /_//_/// /_\ / //_// / //_'/ //     Duration: 0.000     CPU time: 0.000
 /   _/                      v5.0.0
 
-Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\test_validator.py:55
+Profile at G:\UOM\Semester 5\CS3501 - Data Science and Engineering Project\Project\EduGenius\tests\performance_profiling\performance_profile_validator.py:55
 
 No samples were recorded.
